16.01 оболочка правильаня полностью за куб

+++ Multifigures/FIgures/Square.cs
@@ -11,7 +11,7 @@ namespace Multifigures.Figures
 {
     public sealed class Square : Shape
     {
-        public Square(int xx, int yy, Color cc) : base(xx, yy, cc) {}
+        public Square(double xx, double yy, Color cc) : base(xx, yy, cc) {}
 
         public override void Draw(DrawingContext context)
         {
